fix: mutation, query 관련 정리
