--- app/(free-board)/_query/query.ts
@@ -1,120 +1,0 @@
-import {
-  deleteArticlesArticleId,
-  patchArticlesArticleId,
-  postArticles,
-} from "@/lib/apis/article";
-import { showToast } from "@/lib/show-toast";
-import { useMutation } from "@tanstack/react-query";
-import { toast } from "react-toastify";
-
-export function useUploadArticleMutation(close: () => void) {
-  return useMutation({
-    mutationFn: (data: { image: File; title: string; content: string }) =>
-      postArticles({
-        image: data.image,
-        title: data.title,
-        content: data.content,
-      }),
-    onMutate: () => {
-      showToast("loading", "게시글을 등록 중입니다.", {
-        toastId: "uploadPost",
-      });
-      close();
-    },
-    onSuccess: () => {
-      toast.update("uploadPost", {
-        render: "게시글이 성공적으로 등록되었습니다",
-        type: "success",
-        isLoading: false,
-        hideProgressBar: false,
-        autoClose: 1000,
-      });
-    },
-    onError: (error) => {
-      console.error(error);
-      toast.update("uploadPost", {
-        render: "게시글 등록에 실패했습니다.",
-        type: "error",
-        isLoading: false,
-        hideProgressBar: false,
-        autoClose: 1000,
-      });
-    },
-  });
-}
-
-export function useDeleteArticleMutation() {
-  return useMutation({
-    mutationFn: (articleId: number) => deleteArticlesArticleId({ articleId }),
-    onMutate: () => {
-      showToast("loading", "게시글을 삭제 중입니다.", {
-        toastId: "deletePost",
-      });
-    },
-    onSuccess: () => {
-      toast.update("deletePost", {
-        render: "게시글이 성공적으로 삭제되었습니다",
-        type: "success",
-        isLoading: false,
-        hideProgressBar: false,
-        autoClose: 1000,
-      });
-    },
-    onError: (error) => {
-      console.error(error);
-      toast.update("deletePost", {
-        render: error.message,
-        type: "error",
-        isLoading: false,
-        hideProgressBar: false,
-        autoClose: 1000,
-      });
-    },
-  });
-}
-
-export function usePatchArticleMutation() {
-  return useMutation({
-    mutationFn: ({
-      articleId,
-      image,
-      title,
-      content,
-    }: {
-      articleId: number;
-      image: string;
-      title: string;
-      content: string;
-    }) =>
-      patchArticlesArticleId({
-        articleId,
-        image,
-        title,
-        content,
-      }),
-    onMutate: () => {
-      showToast("loading", "게시글을 수정 중입니다.", {
-        toastId: "patchPost",
-      });
-    },
-    onSuccess: () => {
-      toast.update("patchPost", {
-        render: "게시글이 성공적으로 수정되었습니다",
-        type: "success",
-        isLoading: false,
-        hideProgressBar: false,
-        autoClose: 1000,
-      });
-    },
-    onError: (error) => {
-      console.error(error);
-      toast.update("patchPost", {
-        render: error.message,
-        type: "error",
-        isLoading: false,
-        hideProgressBar: false,
-        autoClose: 1000,
-      });
-    },
-  });
-}
